requirements.txt
requests
python-dotenv
ccxt
schedule
fastapi
uvicorn